+++ examples/storybook/stories/Components/ViewerWrapper.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { RichContentWrapper } from 'wix-rich-content-wrapper';
+import { RicosViewer } from 'ricos-viewer';
 import { RichContentViewer } from 'wix-rich-content-viewer';
 import {
   pluginLinkButton,
@@ -79,22 +79,17 @@ const plugins = [
   pluginLinkPreview(),
 ];
 
-const ViewerWrapper = ({ contentState, palette, isMobile, addAnchors, normalize }) => {
+const ViewerWrapper = ({ content, palette, isMobile, addAnchors, normalize }) => {
   const theme = palette ? { theme: 'Palette', palette } : { theme: 'Default' };
-  const RichContentViewerProps = {
-    isMobile,
-    addAnchors,
-    normalize,
-  };
   return (
-    <RichContentWrapper plugins={plugins} {...theme}>
-      <RichContentViewer initialState={contentState} {...RichContentViewerProps} />
-    </RichContentWrapper>
+    <RicosViewer plugins={plugins} {...theme} content={content} isMobile={isMobile}>
+      <RichContentViewer addAnchors={addAnchors} normalize={normalize} />
+    </RicosViewer>
   );
 };
 
 ViewerWrapper.propTypes = {
-  contentState: PropTypes.object,
+  content: PropTypes.object,
   palette: PropTypes.arrayOf(PropTypes.object),
   isMobile: PropTypes.bool,
   addAnchors: PropTypes.oneOfType([PropTypes.bool, PropTypes.string]),