+++ examples/storybook/stories/TestCase/BlogLefties.js
@@ -1,14 +1,16 @@
 import React from 'react';
-import { htmlTypeMapper } from 'wix-rich-content-plugin-html/dist/module.viewer';
+import { pluginHtml } from 'wix-rich-content-plugin-html/dist/module.viewer';
 import { RichContentViewerBox, ContentState, Section, Page } from '../Components/StoryParts';
-import { RichContentViewer } from 'wix-rich-content-viewer';
-import { imageTypeMapper } from 'wix-rich-content-plugin-image/dist/module.viewer';
-import { linkTypeMapper } from 'wix-rich-content-plugin-link/dist/module.viewer';
+import { RicosViewer } from 'ricos-viewer';
+import { pluginImage } from 'wix-rich-content-plugin-image/dist/module.viewer';
+import { pluginGallery } from 'wix-rich-content-plugin-gallery/dist/module.viewer';
+import { pluginVideo } from 'wix-rich-content-plugin-video/dist/module.viewer';
+import { pluginDivider } from 'wix-rich-content-plugin-divider/dist/module.viewer';
 import fixture from '../../../../e2e/tests/fixtures/plugin-left-alignment.json';
 
 import viewerTheme from './viewer.scss';
 
-const typeMappers = [imageTypeMapper, linkTypeMapper, htmlTypeMapper];
+const plugins = [pluginImage(), pluginHtml(), pluginGallery(), pluginVideo(), pluginDivider()];
 const theme = {
   ...viewerTheme,
 };
@@ -18,7 +20,7 @@ export default () => {
     <Page title="HTML Height">
       <Section title={'Height check'}>
         <RichContentViewerBox preset="blog-preset">
-          <RichContentViewer initialState={fixture} typeMappers={typeMappers} theme={theme} />
+          <RicosViewer content={fixture} plugins={plugins} theme={theme} />
         </RichContentViewerBox>
       </Section>