+++ rollup.config.common.js
@@ -48,7 +48,9 @@ export default (output, shouldExtractCss) => {
         input: libEntriesPath + file,
         output: output.map(({ format }) => ({
           format,
-          file: `dist/lib/${format === 'cjs' ? file.replace('.js', '.cjs.js') : file}`,
+          file: `dist/lib/${
+            format === 'cjs' ? file.replace('.js', '.cjs.js').replace('.ts', '.cjs.js') : file
+          }`,
         })),
         plugins,
         external,